chore: move some components around
